+++ backend/cache/BGC.CacheCore/BGC.CacheCore.csproj
@@ -1,17 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
   <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.17.0" />
-    <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.20.1" />
-    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.ServiceBus" Version="5.17.0" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="9.0.10" />
-    <PackageReference Include="Polly" Version="8.6.4" />
+    <PackageReference Include="Azure.Identity"  />
+    <PackageReference Include="Azure.Messaging.ServiceBus"  />    
+    <PackageReference Include="Microsoft.Extensions.Http.Polly"  />
+    <PackageReference Include="Polly"  />
   </ItemGroup>
 
   <ItemGroup>
