--- backend/cache/BGC.Functions/.dockerignore
@@ -1,1 +1,0 @@
-local.settings.json
