--- backend/cache/BGC.Functions/Dockerfile
@@ -1,14 +1,0 @@
-FROM mcr.microsoft.com/dotnet/sdk:7.0 AS installer-env
-
-COPY . /src/dotnet-function-app
-RUN cd /src/dotnet-function-app && \
-mkdir -p /home/site/wwwroot && \
-dotnet publish *.csproj --output /home/site/wwwroot
-
-# To enable ssh & remote debugging on app service change the base image to the one below
-# FROM mcr.microsoft.com/azure-functions/dotnet-isolated:3.0-dotnet-isolated7.0-appservice
-FROM mcr.microsoft.com/azure-functions/dotnet-isolated:4-dotnet-isolated7.0
-ENV AzureWebJobsScriptRoot=/home/site/wwwroot \
-    AzureFunctionsJobHost__Logging__Console__IsEnabled=true
-
-COPY --from=installer-env ["/home/site/wwwroot", "/home/site/wwwroot"]
