--- backend/cache/BGC.Functions/Models/Exceptions/ValidationException.cs
@@ -1,6 +1,0 @@
-﻿namespace BGC.Functions.Models.Exceptions
-{
-    public class ValidationException : Exception
-    {
-    }
-}
