--- backend/cache/BGC.Functions/Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "BGC_CacheQueueFunction": {
-      "commandName": "Project",
-      "commandLineArgs": "--port 7290",
-      "launchBrowser": false
-    }
-  }
-}
