--- backend/cache/BGC.Functions/host.json
@@ -1,47 +1,0 @@
-{
-  "version": "2.0",
-  "extensions": {
-    "serviceBus": {
-      "clientRetryOptions": {
-        "mode": "exponential",
-        "tryTimeout": "00:01:00",
-        "delay": "00:00:00.80",
-        "maxDelay": "00:01:00",
-        "maxRetries": 3
-      },
-      "prefetchCount": 0,
-      "transportType": "amqpWebSockets",
-      "autoCompleteMessages": true,
-      "maxAutoLockRenewalDuration": "00:05:00",
-      "maxConcurrentCalls": 16,
-      "maxConcurrentSessions": 8,
-      "maxMessageBatchSize": 1000,
-      "minMessageBatchSize": 1,
-      "maxBatchWaitTime": "00:00:30",
-      "sessionIdleTimeout": "00:01:00",
-      "enableCrossEntityTransactions": false
-    }
-  },
-  "functionTimeout": "00:00:20",
-  "logging": {
-    "logLevel": {
-      "default": "Warning",
-      "Microsoft": "Warning",
-      "Function": "Error",
-      "Host.Results": "Error",
-      "Host.Aggregator": "Information"
-    },
-    "applicationInsights": {
-      "enablePerformanceCountersCollection": false,
-      "samplingSettings": {
-        "isEnabled": true,
-        "maxTelemetryItemsPerSecond": "10",
-        "excludedTypes": "Request"
-      }
-    },
-    "aggregator": {
-      "batchSize": 1000,
-      "flushTimeout": "00:30:00"
-    }
-  }
-}
