+++ backend/cache/BGC.UpdateBoardGameCacheWorker/BGC.UpdateBoardGameCacheWorker.csproj
@@ -11,13 +11,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.17.0" />
-    <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.20.1" />
-    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.ServiceBus" Version="5.17.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.10" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="9.0.10" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.22.1" />
-    <PackageReference Include="Polly" Version="8.6.4" />
+    <PackageReference Include="Azure.Identity"  />
+    <PackageReference Include="Azure.Messaging.ServiceBus"  />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.ServiceBus"  />
+    <PackageReference Include="Microsoft.Extensions.Hosting"  />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly"  />    
+    <PackageReference Include="Polly"  />
   </ItemGroup>
 
   <ItemGroup>
