--- backend/tests/BGC.Core.UnitTests/Usings.cs
@@ -1,5 +1,0 @@
-global using FluentAssertions;
-global using Microsoft.Extensions.Logging;
-global using MongoDB.Driver;
-global using Moq;
-global using Xunit;
