--- backend/tests/BGC.Functions.UnitTests/Usings.cs
@@ -1,4 +1,0 @@
-global using Xunit;
-global using FluentAssertions;
-global using Microsoft.Extensions.Logging;
-global using Moq;
