--- backend/tests/BGC.SearchApi.UnitTests/Usings.cs
@@ -1,4 +1,0 @@
-global using FluentAssertions;
-global using Microsoft.Extensions.Logging;
-global using Moq;
-global using Xunit;
