+++ backend/tests/BGC.Tests.Core/BGC.Tests.Core.csproj
@@ -1,17 +1,15 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+<Project Sdk="Microsoft.NET.Sdk">   
+  <PropertyGroup>    
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove=".editorconfig" />
+    <PackageReference Include="Moq" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Moq" Version="4.20.72" />
+    <ProjectReference Include="..\BGC.Core\BGC.Core.csproj" />
   </ItemGroup>
 
 </Project>
