+++ backend/tests/BGC.Tests.Core/Extensions/GenericExtensions.cs
@@ -1,11 +1,12 @@
-﻿using System.Text.Json;
-using System.Text;
+﻿using System.Text;
+using System.Text.Json;
 
 namespace BGC.Tests.Core.Extensions
 {
     public static class GenericExtensions
     {
-        public static byte[] ToByteArray<T>(this T model) where T : class
+        public static byte[] ToByteArray<T>(this T model)
+            where T : class
         {
             return Encoding.UTF8.GetBytes(JsonSerializer.Serialize(model));
         }
