+++ backend/tests/BGC.UpdateBoardGameCacheWorker.UnitTests/BGC.UpdateBoardGameCacheWorker.UnitTests.csproj
@@ -6,15 +6,16 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoFixture" />    
-    <PackageReference Include="Azure.Messaging.ServiceBus" />
+    <PackageReference Include="AutoFixture"  />
+    <PackageReference Include="Azure.Identity"  />
+    <PackageReference Include="Azure.Messaging.ServiceBus"  />
     <PackageReference Include="coverlet.msbuild" >
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="FluentAssertions"  />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing"  />
-    <PackageReference Include="Microsoft.NET.Test.Sdk"  />    
+    <PackageReference Include="Microsoft.NET.Test.Sdk"  />
     <PackageReference Include="Moq"  />
     <PackageReference Include="StyleCop.Analyzers" >
       <PrivateAssets>all</PrivateAssets>
@@ -32,8 +33,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\BGC.SearchApi\BGC.SearchApi.csproj" />
-    <ProjectReference Include="..\BGC.Tests.Core\BGC.Tests.Core.csproj" />
-  </ItemGroup>
+    <ProjectReference Include="..\..\cache\BGC.CacheCore\BGC.CacheCore.csproj" />
+    <ProjectReference Include="..\..\cache\BGC.UpdateBoardGameCacheWorker\BGC.UpdateBoardGameCacheWorker.csproj" />
+    <ProjectReference Include="..\BGC.Core.UnitTests\BGC.Core.UnitTests.csproj" />
+  </ItemGroup>  
 
 </Project>
