+++ backend/tests/BGC.UpdateBoardGameCacheWorker.UnitTests/Services/UpdateBoardGameCacheServiceTests.cs
@@ -10,30 +10,30 @@ using BGC.Core.Models.Dtos.BoardGameGeek;
 using BGC.Core.Models.Dtos.BoardGameOracle;
 using BGC.Core.Repositories.Interfaces;
 using BGC.Core.Services.Interfaces;
-using BGC.Functions.Functions;
 using BGC.Tests.Core.Extensions;
+using BGC.UpdateBoardGameCacheWorker.Services;
 
 namespace BGC.Functions.UnitTests.Functions
 {
-    public class UpdateBoardGameCacheFunctionTests
+    public class UpdateBoardGameCacheServiceTests
     {
         private readonly Fixture fixture = new Fixture();
 
-        private readonly Mock<ILogger<UpdateBoardGameCacheFunction>> _mockLogger;
+        private readonly Mock<ILogger<UpdateBoardGameCacheService>> _mockLogger;
         private readonly Mock<IBggService> _mockBggService;
         private readonly Mock<IBoardGamesRepository> _mockBoardGamesRepository;
         private readonly Mock<IBoardGameOracleService> _mockBoardGameOracleService;
 
-        private readonly UpdateBoardGameCacheFunction _updateBoardGameCacheFunction;
+        private readonly UpdateBoardGameCacheService _updateBoardGameCacheService;
 
-        public UpdateBoardGameCacheFunctionTests()
+        public UpdateBoardGameCacheServiceTests()
         {
-            _mockLogger = new Mock<ILogger<UpdateBoardGameCacheFunction>>();
+            _mockLogger = new Mock<ILogger<UpdateBoardGameCacheService>>();
             _mockBggService = new Mock<IBggService>();
             _mockBoardGamesRepository = new Mock<IBoardGamesRepository>();
             _mockBoardGameOracleService = new Mock<IBoardGameOracleService>();
 
-            _updateBoardGameCacheFunction = new UpdateBoardGameCacheFunction(
+            _updateBoardGameCacheService = new UpdateBoardGameCacheService(
                 _mockLogger.Object,
                 _mockBggService.Object,
                 _mockBoardGamesRepository.Object,
@@ -45,7 +45,7 @@ namespace BGC.Functions.UnitTests.Functions
         {
             var mockMessage = ServiceBusModelFactory.ServiceBusReceivedMessage();
 
-            var runFunc = async () => await _updateBoardGameCacheFunction.Run(mockMessage);
+            var runFunc = async () => await _updateBoardGameCacheService.ProcessUpdateCacheMessage(mockMessage);
 
             await runFunc.Should().ThrowAsync<JsonException>();
         }
@@ -61,7 +61,7 @@ namespace BGC.Functions.UnitTests.Functions
             var mockMessage = ServiceBusModelFactory.ServiceBusReceivedMessage(new BinaryData(mockMessageData.ToByteArray()));
             _mockBggService.Setup(service => service.GetDetails(mockBoardGameId.ToString(), It.IsAny<CancellationToken>())).ThrowsAsync(new HttpRequestException());
 
-            var runFunc = async () => await _updateBoardGameCacheFunction.Run(mockMessage);
+            var runFunc = async () => await _updateBoardGameCacheService.ProcessUpdateCacheMessage(mockMessage);
 
             await runFunc.Should().ThrowAsync<HttpRequestException>();
         }
@@ -78,7 +78,7 @@ namespace BGC.Functions.UnitTests.Functions
             _mockBggService.Setup(service => service.GetDetails(mockMessageData.BoardGameId, It.IsAny<CancellationToken>()))
                            .ReturnsAsync(() => mockBoardGameDetails);
 
-            await _updateBoardGameCacheFunction.Run(mockMessage);
+            await _updateBoardGameCacheService.ProcessUpdateCacheMessage(mockMessage);
 
             foreach (RegionDto regionDto in Enum.GetValues(typeof(RegionDto)))
             {
@@ -101,7 +101,7 @@ namespace BGC.Functions.UnitTests.Functions
             _mockBggService.Setup(service => service.GetDetails(mockMessageData.BoardGameId, It.IsAny<CancellationToken>()))
                            .ReturnsAsync(() => mockBoardGameDetails);
 
-            await _updateBoardGameCacheFunction.Run(mockMessage);
+            await _updateBoardGameCacheService.ProcessUpdateCacheMessage(mockMessage);
 
             _mockBoardGamesRepository.Verify(repository => repository.UpsertBoardGame(It.IsAny<BoardGame>(), It.IsAny<CancellationToken>()));
         }
